Refactor helper -> line_helper & date_helper
