+++ bin/save_eps_as_csv.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 import 'models/ean_product.dart';
-import 'utils/helper.dart';
+import 'utils/date_helper.dart';
 import 'utils/shop_selector_helper.dart';
 
 void eanProductListToCSV(List<EANProduct> eanProductList, String csvFilePath,
