+++ bin/specific/s_kaupat/strings_to_products.dart
@@ -1,9 +1,9 @@
-import '../../utils/helper.dart';
+import '../../utils/line_helper.dart';
 import '../../models/receipt_product.dart';
 
 /// Goes through the list of lines and returns a list of products.
 List<ReceiptProduct> strings2Products(List<String> lines) {
-  var helper = Helper();
+  var helper = LineHelper();
   List<ReceiptProduct> products = [];
 
   for (var item in lines) {
